Add files via upload
Problem Statement:
Without analyzing the competition, it is difficult for a business to survive. You are
tasked to analyse the competition for the management to provide better results. This
data set has information on the market capitalization of the top 500 companies in India.
Serial NumberNameName of CompanyMar Cap – CroreMarket Capitalization in
CroresSales Qtr – CroreQuarterly Sale in crores. Find key metrics and factors and
show the meaningful relationships between attributes.
Do your own research and come up with your findings